+++ packages/core/kernal/stories/index.stories.tsx
@@ -6,22 +6,31 @@ import { prefetch } from '../src/prefetch';
 
 start();
 
+const appInfo = {
+  id: 'os-example',
+  manifest: 'https://g.alicdn.com/ConsoleOS/OSExample/0.0.2/os-example.manifest.json'
+}
+
 storiesOf('Basic Console OS', module)
   .add('Basic Use', () => {
     const appRef = useRef();
     const [loading, setLoading] = useState(true);
     useEffect(() => {
-      prefetch([{id: 'os-example', manifest: 'https://g.alicdn.com/ConsoleOS/OSExample/0.0.2/os-example.manifest.json'}]);
+      prefetch([appInfo]);
       setTimeout(() => {
         setLoading(false);
         mountApp({ 
-          id: 'os-example', 
-          manifest: 'https://g.alicdn.com/ConsoleOS/OSExample/0.0.2/os-example.manifest.json',
+          ...appInfo,
           dom: appRef.current,
+        }, {
+          sandBox: {
+            syncInitHref: true,
+            initialPath: '/about'
+          }
         });
       }, 3000)
     }, []);
     return (
-      <div ref={appRef}>{loading && 'l.oading'}</div>
+      <div ref={appRef}>{loading && 'loading...'}</div>
     )
   })