+++ packages/react/react-application/src/index.tsx
@@ -52,6 +52,9 @@ interface IProps<T = any> extends HTMLAttributes<Element> {
 
   disableBodyTag: boolean;
 
+  /**
+   * loading status for consoleos app
+   */
   loading: boolean | React.ReactChild;
 
   appProps: T;
@@ -104,7 +107,7 @@ class Application<T> extends React.Component<Partial<IProps<T>>, IState> {
       const { jsUrl: url, id, manifest, externalsVars, singleton = true } = this.props;
 
       if (!id) {
-        throw Error('You should give a id for OS Application');
+        throw new Error('You should give a id for OS Application');
       }
 
       let sandBox = this.props.sandBox;
@@ -124,9 +127,6 @@ class Application<T> extends React.Component<Partial<IProps<T>>, IState> {
         throw new Error('React dom element is no prepared. please check')
       }
 
-      if (externalsVars) {
-        sandBox.externalsVars = externalsVars;
-      }
       const appInfo = {
         url,
         id,